Add files via upload
Write a connection-oriented client-server program in C where the client will send a number(x) to the server and the server display this number. After that, the server will double the number (2*x) and send it to the client. The client will display the number received from the server. The programs will terminate when 2*x=100.